requirements.txt
alpaca-trade-api==3.0.2
pandas==1.4.4
numpy==1.21.5
yfinance==0.2.12
python-dotenv==0.19.2
ta-lib==0.4.24
textblob==0.15.3
scikit-learn==1.0.2
